# bff-application/src/main/openapi/microservice-api.json
{
  "swagger": "2.0",
  "info": {
    "description": "Example Microservice API",
    "version": "1.0.0",
    "title": "Example Microservice API",
    "contact": {
      "email": "georg.wittberger@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/v1",
  "schemes": [
    "http"
  ],
  "tags": [{
    "name": "messages",
    "description": "Messages API"
  }],
  "paths": {
    "/messages/{id}": {
      "get": {
        "summary": "Returns the message with the given ID",
        "description": "Returns the message with the given ID",
        "operationId": "getMessageById",
        "tags": ["messages"],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the message",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/MessageResponse"
            }
          },
          "500": {
            "description": "Error"
          }
        }
      }
    }
  },
  "definitions": {
    "MessageResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string"
            },
            "id": {
              "type": "string"
            },
            "attributes": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string"
                },
                "lastModified": {
                  "type": "string",
                  "format": "date-time"
                }
              },
              "required": [
                "text",
                "lastModified"
              ]
            }
          },
          "required": [
            "type",
            "id",
            "attributes"
          ]
        }
      },
      "required": [
        "data"
      ]
    }
  }
}